+++ CHANGELOG.md
@@ -10,11 +10,15 @@
 ### Added
 - Command `results` to show test results summary from the command line (CLI equivalent to viewing `results.xml` in a browser).
 - Command `clean` to remove old content of logs directory (previous png screenshots, HTML snapshots and JUnit xmls).
+- Option `--capability` to `run` command which allows to simply pass any custom DesiredCapabilities to the WebDriver server.
+- Support for cloud services like [SauceLabs](https://saucelabs.com/) or [BrowserStack](https://www.browserstack.com/) - simply pass remote server URL (including credentials) using `--server-url`.
+- If [SauceLabs](https://saucelabs.com/) is used as remote cloud platform, test results are automatically published to SauceLabs API when test is done.
 
 ### Changed
 - Content of directory with logs is cleaned before `run` command starts by default. This could be suppressed using `--no-clean` option of the `run` command. ([#63](https://github.com/lmc-eu/steward/issues/63))
 - Directory for logs is automatically created if it does not exist. ([#67](https://github.com/lmc-eu/steward/issues/67))
-- Process output printed to console (when using `-vv` or `-vvv`) is now prefixed with class name
+- Process output printed to console (when using `-vv` or `-vvv`) is now prefixed with class name.
+- BC: Pass instance of current PHPUnit test as third parameter of `publishResult()` method of `AbstractPublisher` descendants.
 
 ### Fixed
 - Parsing of latest Selenium server version in `install` command so that even Selenium 3 beta releases are installed ([#76](https://github.com/lmc-eu/steward/pull/76))